fix linger and tests

+++ src/shadowsocks/server.zig
@@ -319,8 +319,8 @@ fn forwardToClient(state: *ClientState, received: []const u8, allocator: std.mem
 
 fn closeSocketWithRst(socket: network.Socket) void {
     const Linger = extern struct {
-        l_onoff: u16,
-        l_linger: u16,
+        l_onoff: i32,
+        l_linger: i32,
     };
 
     const value: Linger = .{
